Big integer ID

+++ src/models/user.py
@@ -1,13 +1,13 @@
 from .base import Base
 
-from sqlalchemy import String, Integer, Boolean
+from sqlalchemy import String, BigInteger, Boolean
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 
 class User(Base):
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     name: Mapped[str] = mapped_column(String(60), nullable=False)
     date_of_birth_text: Mapped[str] = mapped_column(String(60), nullable=True)
